remove unused modules netutil.py and socks5.py
These files are currently not used anywhere (as can be easily verified via
`git grep netutil` and `git grep socks5`) and unlikely to ever be needed in the
future. The purpose is providing functionalities for testing bitcoind-specific
networking stuff, which is not needed here, hence we can just remove it.