remove unused module wallet_util.py (#26)
This file is currently not used anywhere (as can be easily verified via `git
grep wallet_util`) and unlikely to ever be needed in the future. The purpose is
providing functionalities for testing bitcoind-specific wallet stuff (generating
addresses by communicating with an actual node), which is not needed here, hence
we can just remove it.