remove unneeded class HeaderAndShortIDs (+siphash module)
The class `HeaderAndShortIDs` is currently not used anywhere (as can be
easily verified via `git grep HeaderAndShortIDs`) and unlikely to ever
be needed in the future. The purpose is providing functionalities for
testing the compact block relay (BIP152) protocol, which is not needed
here, hence we can just remove it.

With that, the functions `calculate_shortid` and `siphash256` also
become unused and can be removed as well.